+++ src/main/java/dev/efnilite/ip/menu/community/SingleLeaderboardMenu.java
@@ -21,15 +21,17 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.SkullMeta;
 
-import java.util.*;
-import java.util.stream.Collectors;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 /**
  * Menu for a single leaderboard
  */
 public class SingleLeaderboardMenu {
 
-    public void open(Player player, Mode mode, Sort sort) {
+    public void open(Player player, Mode mode, Leaderboard.Sort sort) {
         Leaderboard leaderboard = mode.getLeaderboard();
 
         if (leaderboard == null) {
@@ -37,19 +39,19 @@ public class SingleLeaderboardMenu {
         }
 
         // init vars
-        ParkourUser user = ParkourUser.getUser(player);
-        String locale = user == null ? Option.OPTIONS_DEFAULTS.get(ParkourOption.LANG) : user.locale;
-        PagedMenu menu = new PagedMenu(3, Locales.getString(player, "%s.name".formatted(ParkourOption.LEADERBOARDS.path)));
+        var user = ParkourUser.getUser(player);
+        var locale = user == null ? Option.OPTIONS_DEFAULTS.get(ParkourOption.LANG) : user.locale;
+        var menu = new PagedMenu(3, Locales.getString(player, "%s.name".formatted(ParkourOption.LEADERBOARDS.path)));
 
-        List<MenuItem> items = new ArrayList<>();
+        var items = new ArrayList<MenuItem>();
 
-        Item base = Locales.getItem(player, "%s.head".formatted(ParkourOption.LEADERBOARDS.path));
+        var base = Locales.getItem(player, "%s.head".formatted(ParkourOption.LEADERBOARDS.path));
 
-        Map<UUID, Score> sorted = sort.sort(leaderboard.scores);
-
-        for (UUID uuid : sorted.keySet()) {
+        for (Map.Entry<UUID, Score> entry : leaderboard.sort(sort).entrySet()) {
             int rank = items.size() + 1;
-            Score score = sorted.get(uuid);
+
+            var uuid = entry.getKey();
+            var score = entry.getValue();
 
             if (score == null) {
                 continue;
@@ -73,7 +75,7 @@ public class SingleLeaderboardMenu {
 
             // if there are more than 36 players, don't show the heads to avoid server crashing
             // and bedrock has no player skull support
-            if (rank <= 36 && !ParkourUser.isBedrockPlayer(player)) {
+            if (rank <= 20 && !ParkourUser.isBedrockPlayer(player)) {
                 OfflinePlayer op = Bukkit.getOfflinePlayer(uuid);
 
                 if (op.getName() != null && !op.getName().startsWith(".")) { // bedrock players' names with geyser start with a .
@@ -107,10 +109,10 @@ public class SingleLeaderboardMenu {
         };
 
         // get next sorting type
-        Sort next = switch (sort) {
-            case SCORE -> Sort.TIME;
-            case TIME -> Sort.DIFFICULTY;
-            default -> Sort.SCORE;
+        var next = switch (sort) {
+            case SCORE -> Leaderboard.Sort.TIME;
+            case TIME -> Leaderboard.Sort.DIFFICULTY;
+            default -> Leaderboard.Sort.SCORE;
         };
 
         menu.displayRows(0, 1)
@@ -132,42 +134,5 @@ public class SingleLeaderboardMenu {
                     Menus.LEADERBOARDS.open(event.getPlayer());
                 }))
                 .open(player);
-    }
-
-    public enum Sort {
-
-        SCORE {
-            @Override
-            Map<UUID, Score> sort(Map<UUID, Score> scores) {
-                return scores; // already sorted
-            }
-        }, TIME {
-            @Override
-            Map<UUID, Score> sort(Map<UUID, Score> scores) {
-                return scores.entrySet().stream().sorted((o1, o2) -> {
-                            long one = o1.getValue().getTimeMillis();
-                            long two = o2.getValue().getTimeMillis();
-
-                            return Math.toIntExact(one - two); // natural order (lower == better)
-                        }) // reverse natural order
-                        .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (a, b) -> b, LinkedHashMap::new));
-            }
-        }, DIFFICULTY {
-            @Override
-            Map<UUID, Score> sort(Map<UUID, Score> scores) {
-                return scores.entrySet().stream().sorted((o1, o2) -> {
-                            String first = o1.getValue().difficulty();
-                            String second = o2.getValue().difficulty();
-
-                            double one = Double.parseDouble(first.equals("?") ? "1.0" : first);
-                            double two = Double.parseDouble(second.equals("?") ? "1.0" : second);
-
-                            return (int) (100 * (two - one)); // reverse natural order (higher == better)
-                        }) // reverse natural order
-                        .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (a, b) -> b, LinkedHashMap::new));
-            }
-        };
-
-        abstract Map<UUID, Score> sort(Map<UUID, Score> scores);
     }
 }