+++ src/main/java/dev/efnilite/ip/player/data/PreviousData.java
@@ -3,12 +3,13 @@ package dev.efnilite.ip.player.data;
 import dev.efnilite.ip.IP;
 import dev.efnilite.ip.config.Config;
 import dev.efnilite.ip.config.Option;
-import dev.efnilite.ip.reward.RewardString;
+import dev.efnilite.ip.reward.Reward;
 import io.papermc.lib.PaperLib;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.bukkit.potion.PotionEffect;
+import org.bukkit.util.Vector;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
@@ -30,9 +31,9 @@ public class PreviousData {
     private final Collection<PotionEffect> effects;
 
     /**
-     * List of all {@link RewardString} to execute on leave.
+     * List of all {@link Reward} to execute on leave.
      */
-    public List<RewardString> onLeave = new ArrayList<>();
+    public List<Reward> onLeave = new ArrayList<>();
 
     public PreviousData(@NotNull Player player) {
         gamemode = player.getGameMode();
@@ -79,6 +80,11 @@ public class PreviousData {
             for (PotionEffect effect : effects) {
                 player.addPotionEffect(effect);
             }
+
+            player.resetPlayerTime();
+            player.resetPlayerWeather();
+            player.setVelocity(new Vector(0, 0, 0));
+            player.setFallDistance(0f);
         } catch (Exception ex) { // not the best way to do this... too bad!
             IP.logging().stack("Error while recovering stats of %s".formatted(player.getName()), ex);
         }