+++ src/main/java/dev/efnilite/ip/world/WorldManager.java
@@ -2,31 +2,27 @@ package dev.efnilite.ip.world;
 
 import dev.efnilite.ip.IP;
 import dev.efnilite.ip.config.Config;
-import dev.efnilite.ip.config.Option;
-import org.bukkit.Bukkit;
-import org.bukkit.Difficulty;
-import org.bukkit.GameRule;
-import org.bukkit.World;
+import dev.efnilite.vilib.util.VoidGenerator;
+import org.bukkit.*;
 
-public interface WorldManager {
+import java.io.File;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Comparator;
+import java.util.stream.Stream;
 
-    /**
-     * Implementation for creating a world.
-     *
-     * @return The created world.
-     */
-    World createWorld();
+public class WorldManager {
 
-    /**
-     * Implementation for deleting a world.
-     */
-    void deleteWorld();
+    private static String name;
+    private static World world;
 
     /**
      * Creates a new world and sets all according settings in it.
      */
-    static void create() {
-        World world = getWorld();
+    public static void create() {
+        name = Config.CONFIG.getString("world.name");
+
+        var world = Bukkit.getWorld(name);
 
         if (!Config.CONFIG.getBoolean("joining")) {
             return;
@@ -36,15 +32,32 @@ public interface WorldManager {
             IP.logging().warn("Crash detected! The parkour world loading twice is not usual behaviour. This only happens after a server crash.");
         }
 
-        WorldManager manager = getInstance();
-
-        IP.log("Initializing world rules");
-
         if (Config.CONFIG.getBoolean("world.delete-on-reload")) {
-            manager.deleteWorld();
+            deleteWorld();
         }
 
-        world = manager.createWorld();
+        createWorld();
+        setup();
+    }
+
+    private static void createWorld() {
+        IP.log("Creating Spigot world");
+
+        try {
+            WorldCreator creator = new WorldCreator(name)
+                    .generateStructures(false)
+                    .type(WorldType.NORMAL)
+                    .generator(VoidGenerator.getGenerator()) // to fix No keys in MapLayer etc.
+                    .environment(World.Environment.NORMAL);
+
+            world = Bukkit.createWorld(creator);
+        } catch (Exception ex) {
+            IP.logging().stack("Error while trying to create the parkour world", "delete the parkour world folder and restart the server", ex);
+        }
+    }
+
+    private static void setup() {
+        IP.log("Initializing world rules");
 
         world.setGameRule(GameRule.DO_FIRE_TICK, false);
         world.setGameRule(GameRule.DO_MOB_SPAWNING, false);
@@ -65,7 +78,7 @@ public interface WorldManager {
     /**
      * Deletes the world.
      */
-    static void delete() {
+    public static void delete() {
         if (!Config.CONFIG.getBoolean("world.delete-on-reload") || !Config.CONFIG.getBoolean("joining")) {
             return;
         }
@@ -73,24 +86,41 @@ public interface WorldManager {
 
         getWorld().getPlayers().forEach(player -> player.kickPlayer("Server is restarting"));
 
-        getInstance().deleteWorld();
+        deleteWorld();
+    }
+
+    private static void deleteWorld() {
+        IP.log("Deleting Spigot world");
+
+        File file = new File(name);
+
+        // world has already been deleted
+        if (!file.exists()) {
+            return;
+        }
+
+        Bukkit.unloadWorld(name, false);
+
+        try (Stream<Path> files = Files.walk(file.toPath())) {
+            files.sorted(Comparator.reverseOrder())
+                    .map(Path::toFile)
+                    .forEach(File::delete);
+        } catch (Exception ex) {
+            IP.logging().stack("Error while trying to delete the parkour world", ex);
+        }
     }
 
     /**
-     * Gets the parkour world.
-     *
+     * @return the name of the parkour world.
+     */
+    public static String getName() {
+        return name;
+    }
+
+    /**
      * @return the Bukkit world wherein IP is currently active.
      */
-    static World getWorld() {
-        return Bukkit.getWorld(Option.WORLD_NAME);
-    }
-
-    /**
-     * Returns the appropriate instance.
-     *
-     * @return The appropriate instance.
-     */
-    static WorldManager getInstance() {
-        return WorldManagerMV.MANAGER != null ? new WorldManagerMV() : new WorldManagerMC();
+    public static World getWorld() {
+        return world;
     }
 }