+++ CHANGES.md
@@ -1,4 +1,7 @@
 **[Infinite Elytra Parkour](https://www.spigotmc.org/resources/115322/) | [Infinite Parkour+](https://www.spigotmc.org/resources/105019/)**
 
-- Fixed first release of major versions not working
-- Fixed spectator permission not working with /ip join
+- Running Multiverse with IP now does not require VoidGen
+- Improved responsiveness of parkour and scoreboard
+- Fixed leaderboards sorting being wrong
+- Fixed placeholders returning wrong values
+- Fixed vague sorting locale in leaderboard menu